+++ src/common/sub-parse/tile-set.js
@@ -1,8 +1,5 @@
 // Grid tiles.
 
-const entryWalls = /[0]/g;					// Zero Regex (Wall)
-const entryFloor = /[1]/g;					// One Regex (Floor)
-
 const wallTileChar = "X";					// Wall
 const floorTileChar = ".";					// Floor
 const nodeTileChar = "O";					// Node
@@ -14,8 +11,8 @@ function swapEntryTileCharacters(rString)
 {
 	var swapRes = rString;
 	
-	swapRes = swapRes.replace(entryWalls, wallTileChar);
-	swapRes = swapRes.replace(entryFloor, floorTileChar);
+	swapRes = swapRes.replace(/[0]/g, wallTileChar);
+	swapRes = swapRes.replace(/[1]/g, floorTileChar);
 	
 	return swapRes;
 }