+++ changelog.md
@@ -1,6 +1,24 @@
 # Changelog
 
-* Rewrote the following files as 'readme.md' in their respective folders:
-	* ./result-folder/context.txt
-	* ./src/common/precompiled-graph/context.txt
-	* ./src/common/sub-interface/spin-text/context.txt
+**./src/common/sub-files/stream-exceptions.js**
+* Corrected typo in 'getFolderCreateErrorText'
+	* 'prefixRegex' to 'prefixSyntax'
+
+---
+
+**./src/common/sub-parse/line-types.js**
+* Removed 'numberBegin' RegEx.
+	* Now used locally in 'checkNumberLineType'
+
+---
+
+**./src/common/sub-parse/tile-set.js**
+* Removed RegEx variables
+	* entryWalls
+	* entryFloor
+* RegEx are now used locally in 'swapEntryTileCharacters'
+
+---
+
+**./src/common/value-prep.js**
+* Removed "Captures whitespace." comment.